Recombine log and linear scaling behavior into one FrequencyAxisLayer component
