+++ ui/src/components/Bouts/FrequencyAxisLayer.tsx
@@ -6,33 +6,67 @@ import { TICKER_HEIGHT } from "./SpectrogramTimeline";
 
 const MAX_TICK_WIDTH = 0.3;
 
-export function FrequencyAxisLinearLayer({
+type Scaling = "log" | "linear";
+
+export function FrequencyAxisLayer({
+  scaling,
   minFrequency,
   maxFrequency,
   zIndex,
 }: {
+  scaling: Scaling;
   minFrequency: number;
   maxFrequency: number;
   zIndex: number;
 }) {
-  const maxScale = Math.ceil(Math.log10(maxFrequency));
-  const minScale = maxScale - 2;
+  let ticks;
+  if (scaling === "log") {
+    const adjustedMinFrequency = Math.max(1, minFrequency);
+    const maxScale = Math.ceil(Math.log10(maxFrequency));
+    const minScale = Math.floor(Math.log10(adjustedMinFrequency));
+    ticks = _.range(minScale, maxScale).flatMap((scale) => {
+      return _.range(
+        Math.pow(10, scale),
+        Math.pow(10, scale + 1),
+        Math.pow(10, scale),
+      )
+        .map((frequency) => {
+          const bottom = `${(100 * (Math.log10(frequency) - Math.log10(adjustedMinFrequency))) / (Math.log10(maxFrequency) - Math.log10(adjustedMinFrequency))}%`;
+          const minWidth = 0.2;
+          const isMajor = frequency === Math.pow(10, scale);
+          const widthFactor = isMajor
+            ? 1
+            : (1 - minWidth) * (frequency / Math.pow(10, scale + 1)) + minWidth;
 
-  const minTick =
-    Math.floor(minFrequency / Math.pow(10, minScale)) * Math.pow(10, minScale);
-  const maxTick =
-    Math.ceil(maxFrequency / Math.pow(10, minScale)) * Math.pow(10, minScale);
+          return { frequency, bottom, widthFactor, isMajor };
+        })
+        .filter(
+          ({ frequency }) =>
+            frequency >= adjustedMinFrequency && frequency <= maxFrequency,
+        );
+    });
+  } else {
+    const maxScale = Math.ceil(Math.log10(maxFrequency));
+    const minScale = maxScale - 2;
 
-  const ticks = _.range(minTick, maxTick, Math.pow(10, minScale))
-    .map((frequency) => {
-      const bottom = `${(100 * frequency) / maxTick}%`;
-      const minWidth = 0.2;
-      const isMajor = frequency % Math.pow(10, minScale) === 0;
-      const widthFactor = isMajor ? 1 : 1 - minWidth;
+    const minTick =
+      Math.floor(minFrequency / Math.pow(10, minScale)) *
+      Math.pow(10, minScale);
+    const maxTick =
+      Math.ceil(maxFrequency / Math.pow(10, minScale)) * Math.pow(10, minScale);
 
-      return { frequency, bottom, widthFactor, isMajor };
-    })
-    .filter(({ frequency }) => frequency > minTick && frequency <= maxTick);
+    ticks = _.range(minTick, maxTick, Math.pow(10, minScale))
+      .map((frequency) => {
+        const bottom = `${(100 * frequency) / maxTick}%`;
+        const minWidth = 0.2;
+        const isMajor = frequency % Math.pow(10, minScale) === 0;
+        const widthFactor = isMajor ? 1 : 1 - minWidth;
+
+        return { frequency, bottom, widthFactor, isMajor };
+      })
+      .filter(({ frequency }) => frequency > minTick && frequency <= maxTick);
+  }
+
   return (
     <Box
       data-component="freq"
@@ -54,7 +88,13 @@ export function FrequencyAxisLinearLayer({
       >
         <Typography
           fontSize={10}
-          sx={{ position: "absolute", top: 0, right: "3px" }}
+          sx={{
+            position: "absolute",
+            top: 0,
+            ...(scaling === "linear"
+              ? { right: "3px" }
+              : { marginLeft: "5px" }),
+          }}
         >
           Hz
         </Typography>
